requirements.txt
Flask==2.3.2
Flask-SocketIO==5.5.1
python-dotenv==1.0.0
pymongo==4.4.1
openai==0.27.0
